OFString: Rework number parsing API
This solves the old signed vs. unsigned problem and allows for more
bases than just 8, 10 and 16, as well as auto-detection of the base (if
base is 0).

FossilOrigin-Name: b6ee372b98879af381829d2ead95ec8fad06a45c31a5dbd626c159026ee10d63

+++ src/OFINICategory.h
@@ -81,8 +81,8 @@ OF_SUBCLASSING_RESTRICTED
  * @return The integer value for the specified key or the specified default
  *	   value if it does not exist
  */
-- (intmax_t)integerForKey: (OFString *)key
-	     defaultValue: (intmax_t)defaultValue;
+- (long long)integerForKey: (OFString *)key
+	      defaultValue: (long long)defaultValue;
 
 /*!
  * @brief Returns the bool value for the specified key or the specified default
@@ -163,7 +163,7 @@ OF_SUBCLASSING_RESTRICTED
  * @param integer The integer to which the value of the key should be set
  * @param key The key for which the new value should be set
  */
-- (void)setInteger: (intmax_t)integer
+- (void)setInteger: (long long)integer
 	    forKey: (OFString *)key;
 
 /*!
